API Integration/llm-api-pipeline/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
numpy>=1.21.0
pandas
matplotlib
seaborn
scikit-learn
